dom/tests/mochitest/pointerlock/test_MozPointerLock.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Test for Bug 633602</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="start();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
  <p id="display"></p>

  <div id="content" style="display: none">
  </div>

  <pre id="test">
    <script class="testbody" type="text/javascript">
    function start() {
      SimpleTest.waitForExplicitFinish();
      var pointer = navigator.mozPointer;
      ok(pointer, "expected navigator.pointer to exist.");
      ok(pointer instanceof MozPointerLock,
         "expected navigator.pointer to be a MozPointerLock.");
      ok(MozPointerLock.prototype.lock, "MozPointerLock should have lock!.");
      ok(MozPointerLock.prototype.unlock,
         "MozPointerLock should have unlock.");
      ok(MozPointerLock.prototype.hasOwnProperty("isLocked"),
         "MozPointerLock should have islocked.");
      is(pointer.isLocked, false, "Mouse Lock should initially be false.");
      SimpleTest.finish();
    }
    </script>
  </pre>
</body>
</html>
